+++ src/main/java/com/example/swp391/repository/StaffProjectRepository.java
@@ -2,8 +2,7 @@ package com.example.swp391.repository;
 
 import com.example.swp391.entity.*;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
+
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -12,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface StaffProjectRepository extends JpaRepository<StaffProjectEntity, Integer> {
     StaffProjectEntity findTopByOrderByStaffProjectIDDesc();
-    List<StaffProjectEntity> findByProject_ProjectID(Integer projectID);
+    List<StaffProjectEntity> findByProject_ProjectID(Long projectID);
     List<StaffProjectEntity> findByStatus(String status);
 
     List<StaffProjectEntity> findByStaff_StaffID(int staffId);
